// src/api/models/couponListItem.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * Allta Admin
 * Allta Admin 서버 문서입니다.
 * OpenAPI spec version: 1.0
 */
import type { CouponUserItem } from './couponUserItem';
import type { CouponStoreItem } from './couponStoreItem';

export interface CouponListItem {
  /** ID */
  id: string;
  /** 쿠폰 이름 */
  name: string;
  /** 쿠폰 종류 */
  type: string;
  /**
   * 이용권 종류
   * @nullable
   */
  passType?: string | null;
  /**
   * 서비스 종류
   * @nullable
   */
  serviceType?: string | null;
  /** 할인 종류 */
  discountType: string;
  /** 할인 금액 */
  discountValue: number;
  /** 생성일 */
  createdAt: string;
  /** 만료일 */
  expiredAt: string;
  /**
   * 사용일
   * @nullable
   */
  usedAt?: string | null;
  /** 사용 여부 */
  isUsed: boolean;
  /** 회원 정보 */
  user: CouponUserItem;
  /** 쿠폰 적용 매장 목록 */
  stores: CouponStoreItem[];
}